+++ src/core/ObjectiveFunctionValidator.ts
@@ -1,13 +1,11 @@
-class ObjectiveFunctionValidator {
-    static validate(objectiveFunction: string): boolean {
-        // Eliminar todos los espacios y luego trim para quitar espacios al principio y al final
-        const cleanedObjectiveFunction = objectiveFunction.replace(/\s+/g, '').trim();
+    class ObjectiveFunctionValidator {
+        static validate(objectiveFunction: string): boolean {
+            const cleanedObjectiveFunction = objectiveFunction.replace(/\s+/g, '').trim();
 
-        // Patrón para validar la función objetivo
-        const pattern = /^(max|min)\w=(([+-]?\d*\.?\d*\w\d+)([+-]\d*\.?\d*\w\d+)*)$/;
+            const pattern = /^(max|min)\w=(([+-]?\d*\.?\d*\w\d+)([+-]\d*\.?\d*\w\d+)*)$/;
 
-        return pattern.test(cleanedObjectiveFunction);
+            return pattern.test(cleanedObjectiveFunction);
+        }
     }
-}
 
-export default ObjectiveFunctionValidator;
+    export default ObjectiveFunctionValidator;
